trivial refactoring to make the frontend cleaner
