+++ frontend/src/routes/ChildRoutes.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import { Redirect, Route, Switch, useHistory } from 'react-router-dom';
-
-import { useAuth } from '../hooks/useAuth';
-import { BoardRoute, DocListRoute, MasterWindowRoute } from './KeyRoutes';
-
+import { Route, Switch } from 'react-router-dom';
+import MasterWindowRoute from './MasterWindowRoute';
+import DocumentViewRoute from './DocumentViewRoute';
+import BoardRoute from './BoardRoute';
 import Board from '../containers/Board.js';
 import Dashboard from '../containers/Dashboard.js';
 import InboxAll from '../containers/InboxAll.js';
 import NavigationTree from '../containers/NavigationTree.js';
 import CalendarPage from '../pages/calendar/CalendarPage';
+import LogoutRoute from './LogoutRoute';
 
 /**
  * @file Functional component.
@@ -16,36 +16,17 @@ import CalendarPage from '../pages/calendar/CalendarPage';
  * Wrapper around the routes that need to be authentication-protected
  */
 const ChildRoutes = () => {
-  const auth = useAuth();
-  const history = useHistory();
-  const loggedIn = auth.isLoggedIn;
-
   return (
     <>
       <Switch>
         <Route exact path="/" component={Dashboard} />
-        <Route
-          path="/window/:windowId/:docId"
-          render={(params) => <MasterWindowRoute {...params} />}
-        />
-        <Route path="/window/:windowId" component={DocListRoute} />
+        <Route path="/window/:windowId/:docId" component={MasterWindowRoute} />
+        <Route path="/window/:windowId" component={DocumentViewRoute} />
         <Route path="/sitemap" component={NavigationTree} />
         <Route path="/board/:boardId" component={BoardRoute} />
         <Route path="/inbox" component={InboxAll} />
         <Route path="/calendar" component={CalendarPage} />
-        <Route
-          path="/logout"
-          render={() => {
-            if (loggedIn) {
-              auth.logout().finally(() => {
-                history.push('/login');
-                return null;
-              });
-            } else {
-              return <Redirect to="/login" />;
-            }
-          }}
-        />
+        <Route path="/logout" component={LogoutRoute} />
         <Route path="*" render={(props) => <Board board="404" {...props} />} />
       </Switch>
     </>
